refactor: 코드 리팩토링
- 불명확한 함수명 변경(checkNameError -> hasNoNameError 등)
- addContactViewController extension 별도의 파일로 분리
- addContactViewController 내 코드를 properties, functions에 맞게 이동
- saveButtonTapped() 내 많은 코드 함수로 분리
